Add : useShuffledData 커스텀 훅을 만들어 분리한다
